Extract useAutoCollapsingMeasuredView into separate hook
